requirements_dev.txt
pip==22.3.1
wheel==0.38.4
asyncio==3.4.3
trio==0.22.0
anyio==3.6.2
twine==4.0.2
sphinx==6.2.1
pytest==7.3.1
pytest-cov==4.0.0
coveralls==3.3.1
sphinx-rtd-theme==1.2.0
asks==3.0.0
python-dotenv==1.0.0